Make only non-human data public automatically
Ensure that the scripted process of making data automatically public
only works for data that is non-human - this is because human data has
legal issues with privacy, security and the like.

Human data is made public only via a manual process where (a) human(s)
get to make a judgement call.