JWT refresh: Deactivate the checks and revocation
The checks for whether a token is already linked, and then revoking it
and raising an error were causing issues in multi-threaded
environments, where there'd be multiple requests to the auth server
all using an expired token.

This just links the refresh token and avoids the check and revocation
for the time being.