Reduce UI complexity
Providing both the "Enter Verification Token" and the "Send
Verification Email" elements of the same user interface seems to
confuse users.

This commit ensures that the system will provide one or the other, but
not both, depending on whether or not there is a pending verification
token present for the user.